Fixing wrong data type

+++ src/app.py
@@ -54,7 +54,7 @@ class OpenSourceCombatlogReader(OscrGui):
         Prepares settings.
         """
         _, _, screen_width, _ = self.app.primaryScreen().availableGeometry().getRect()
-        self.settings['sidebar_item_width'] = self.theme['s.c']['sidebar_item_width'] * screen_width
+        self.settings['sidebar_item_width'] = int(self.theme['s.c']['sidebar_item_width'] * screen_width)
         self.settings['base_path'] = self.format_path(self.app_dir)
         style_path = rf"{self.app_dir}\\{self.settings['plot_stylesheet_path']}"
         self.settings['plot_stylesheet_path'] = os.path.abspath(style_path)
